refactor: extract Dropzone component and useMeetingUpload hook from UploadMeeting
